Broke code to small scripts
