+++ extracting-frames-from-video/extract_frames.py
@@ -1,5 +1,4 @@
 import cv2
-import numpy as np
 from enum import Enum
 
 
@@ -80,16 +79,3 @@ def extract_frames_from_video(video_name):
         list_of_frames = extract_frames_by_number(cap, number_of_frames)
 
     return list_of_frames
-
-
-def make_sift(list_of_frames):
-    list_of_sifted_frames = []
-    for i in list_of_frames:
-        list_of_sifted_frames += cv2.SIFT(i)
-    return list_of_sifted_frames
-
-
-def main():
-    video_name = input("Enter name of video: ")
-    list_of_frames = extract_frames_from_video(video_name)
-    list_of_keypoints = make_sift(list_of_frames)
