+++ extracting-frames-from-video/extract_frames_test.py
@@ -1,0 +1,6 @@
+import unittest
+
+# some tests here
+
+if __name__ == '__main__':
+    unittest.main()
